Use a single state object for the state machine.
The old version were based on the principle of having each state represented by a class, for example : INITIAL, LOADING and LOADED.
This new version is based on the idea of an "evolving" single class representing the state of the machine.
This means that on each transition the properties of the same object are updated, and the new state is represented by the updated values.